Make bower-able
Adding this file allows us to use bower to control the versioning of this package, rather than using the CDNified versions and hoping that they won't conflict. This file is automatically updated by Travis to point at the CDN-ified script that will be generated by frau-ci in the next step. Since the new version isn't released until _after_ this commit is pushed, it does theoretically mean that if somebody pulled master at just the right moment, d2l-fetch-dedupe.html would be pointing to a non-existant file, but oh well. (If the consumer is using a proper bower version, this situation can't occur since the new version is only generated after the new file is CDN-ified.)

This also removes frau-ci, as it has some issues with this approach; instead, we more or less accomplish the same steps with the shell script.